Fix devcard build --- cljc forbidden
It seems like a good idea to have your spec files be in cljc, so that they can
be used wherever. But, evidently cljc was interacting strangely with devcards.
Probably what you'd have to do is exclude the devcards related code from clj
compilation. But for the moment I'm just using cljs files instead of cljc files,
and it works.